+++ src/components/atoms/index.ts
@@ -1,6 +1,6 @@
 // src/components/atoms/index.ts
 
-export { Button } from './Button'
+export { Button } from './Button/Button'
 export { Input } from './Input'
 export { Text, Heading, Price } from './Text'
 export { Badge } from './Badge'
